[NUI] Introduce CornerRadiusPolicy to View
The CornerRadiusPolicy indicates whether the CornerRadius value is relative or absolute.
The default policy is absolute, which means the CornerRadius is an absolute float length.
If the policy is relative, the CornerRadius can have percentage value between 0.0f to 1.0f,
that is relative to the smaller of view's width and height.

Signed-off-by: Jiyun Yang <ji.yang@samsung.com>