[NUI] fix ViewItem can be bindable and propagate invoke BindingConextedChanged
currently binding SetValue is only enabled for Xaml created object.
To make it works for manual binding, make IsCreatedByXaml value true in
constructor as workaround fix.

BindingConextedChanged event is not propagate properly as Elements only
handles LogicalChildrenInternal not Children,
so ViewItem manually propgate event to it's children for workaround fix.

WARNING : this fix must be changed in upper class level not ViewItem!